requirements.txt
aiohttp==3.7.4
altgraph==0.17
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
chardet==3.0.4
click==8.0.1
cloudscraper==1.2.58
colorama==0.4.4
cssselect==1.1.0
cx-Freeze==6.7
cx-Logging==3.0
deepdiff==5.5.0
discord==1.7.3
discord.py==1.7.3
fake-useragent==0.1.11
Faker==8.10.3
Flask==2.0.1
free-proxy==1.0.2
future==0.18.2
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.21.0
oauthlib==3.1.1
ordered-set==4.0.2
packaging==21.0
pandas==1.3.0
parse==1.19.0
pefile==2021.5.24
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyee==8.1.0
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2021.1
random-user-agent==1.0.1
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
schedule==1.1.0
selenium==3.141.0
six==1.15.0
soupsieve==2.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.61.2
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.5
w3lib==1.22.0
websockets==8.1
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0
